Stop bot movement when it has no probes left
The target location of the bot is updated whenever one of the players
probes and adds a tangent to the game. Prior to this patch, the bot
would still move to the new location, even if had no probes left.
With this patch, the fake input generation for the bot stops as soon as
it has used up all its probes.